renamed and start new race
