+++ lib/models/race.js
@@ -33,7 +33,7 @@ raceSchema.statics.RaceScheduler = function(){
     distance = (Math.floor(Math.random() * (10000 -100 +1)) +100)*12;
     prize = distance / 6;
     return {
-        planetArray: this.planet,//TODO: get 1 random planet from the array
+        planet: this.planet,//TODO: get 1 random planet form sample aggregation
         distance: distance, 
         prize: prize
     };
