+++ lib/scripts/create-race.js
@@ -1,0 +1,10 @@
+const Race = require('../models/race');
+
+
+
+module.exports = function() {
+    new Race(raceSchema.RaceScheduler(), raceInterval){
+
+    }
+
+};
