--- lib/scripts/create.js
@@ -1,1 +1,0 @@
-//TODO: export "check for completed races" function
