+++ lib/scripts/race-runner.js
