+++ lib/scripts/race-scheduler.js
@@ -1,5 +1,5 @@
-const checkPlaceholder = require('./check');
-const createPlaceholder = require('./create');
+const checkPlaceholder = require('./race-runner');
+const createPlaceholder = require('./create-race');
 
 module.exports = function() {
