--- test/e2e/create.test.js
@@ -1,27 +1,0 @@
-// const { assert } = require('chai');
-// const db = require('./db');
-// const schedule = require('../../lib/timers/create');
-// const Race = require('../../lib/models/race');
-
-
-
-
-// describe('Auth test', () => {
-
-//     beforeEach(() => {
-//         db.drop();
-
-//     });
-
-//     //TODO: finish test once race schema is completed
-//     it('should create races', () => {
-//         return schedule()
-//             .then( () => {
-//                 return Race.find().lean();
-//             })
-//             .then( races => {
-//                 assert
-//             });
-
-//     });
-// });
